Fixes: removed response message on status code 200 and moved db file under db directory
